requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
